+++ Sources/NextcloudKit/Models/Assistant/v2/TaskList.swift
@@ -4,51 +4,6 @@
 
 import SwiftyJSON
 
-//public class NKTextProcessingTaskV2 {
-//    public var id: Int?
-//    public var type: String?
-//    public var status: Int?
-//    public var userId: String?
-//    public var appId: String?
-//    public var input: String?
-//    public var output: String?
-//    public var identifier: String?
-//    public var completionExpectedAt: Double?
-//
-//    public init(id: Int? = nil, type: String? = nil, status: Int? = nil, userId: String? = nil, appId: String? = nil, input: String? = nil, output: String? = nil, identifier: String? = nil, completionExpectedAt: Double? = nil) {
-//        self.id = id
-//        self.type = type
-//        self.status = status
-//        self.userId = userId
-//        self.appId = appId
-//        self.input = input
-//        self.output = output
-//        self.identifier = identifier
-//        self.completionExpectedAt = completionExpectedAt
-//    }
-//
-//    public init?(json: JSON) {
-//        self.id = json["id"].int
-//        self.type = json["type"].string
-//        self.status = json["status"].int
-//        self.userId = json["userId"].string
-//        self.appId = json["appId"].string
-//        self.input = json["input"].string
-//        self.output = json["output"].string
-//        self.identifier = json["identifier"].string
-//        self.completionExpectedAt = json["completionExpectedAt"].double
-//    }
-//
-//    static func factories(data: JSON) -> [NKTextProcessingTask]? {
-//        guard let allResults = data.array else { return nil }
-//        return allResults.compactMap(NKTextProcessingTask.init)
-//    }
-//
-//    static func factory(data: JSON) -> NKTextProcessingTask? {
-//        NKTextProcessingTask(json: data)
-//    }
-//}
-
 public struct TaskList: Codable {
     public var tasks: [AssistantTask]
